+++ tests/test_errors.py
@@ -1,5 +1,12 @@
 # -*- coding: utf-8 -*-
 
+from wemake_python_styleguide.errors.base import BaseStyleViolation
+
+
+def test_checker_default_location():
+    """Ensures that `BaseStyleViolation` returns correct location."""
+    assert BaseStyleViolation(None)._location() == (0, 0)
+
 
 def test_all_unique_error_codes(all_errors):
     """Ensures that all errors have unique error codes."""