+++ tests/test_visitors/test_ast/conftest.py
@@ -1,0 +1,35 @@
+# -*- coding: utf-8 -*-
+
+import ast
+from textwrap import dedent
+
+import pytest
+
+
+def _maybe_set_parent(tree: ast.AST) -> ast.AST:
+    """
+    Sets parents for all nodes that do not have this prop.
+
+    This step is required due to how `flake8` works.
+    It does not set the same properties as `ast` module.
+
+    This function was the cause of `issue-112`.
+
+    .. versionchanged:: 0.0.11
+
+    """
+    for statement in ast.walk(tree):
+        for child in ast.iter_child_nodes(statement):
+            if not hasattr(child, 'parent'):  # noqa: Z112
+                setattr(child, 'parent', statement)
+
+    return tree
+
+
+@pytest.fixture(scope='session')
+def parse_ast_tree():
+    """Helper function to convert code to ast."""
+    def factory(code: str) -> ast.AST:
+        return _maybe_set_parent(ast.parse(dedent(code)))
+
+    return factory