+++ wemake_python_styleguide/errors/base.py
@@ -1,7 +1,14 @@
 # -*- coding: utf-8 -*-
 
 import ast
-from typing import Optional, Tuple
+import tokenize
+from typing import Tuple, Union
+
+ErrorNode = Union[
+    ast.AST,
+    tokenize.TokenInfo,
+    None,
+]
 
 
 class BaseStyleViolation(object):
@@ -18,8 +25,8 @@ class BaseStyleViolation(object):
     code: int
     should_use_text: bool = True
 
-    def __init__(self, node: Optional[ast.AST], text: str = None) -> None:
-        """Creates new instance of AST style violation."""
+    def __init__(self, node: ErrorNode, text: str = None) -> None:
+        """Creates new instance of style violation."""
         self._node = node
 
         if text is None:
@@ -28,8 +35,11 @@ class BaseStyleViolation(object):
             self._text = text
 
     def _full_code(self) -> str:
-        """Returns fully formated code."""
-        return 'Z' + str(self.code)
+        """Returns fully formatted code."""
+        return 'Z' + str(self.code).zfill(3)
+
+    def _location(self) -> Tuple[int, int]:
+        return 0, 0
 
     def message(self) -> str:
         """Returns error's formatted message."""
@@ -39,22 +49,32 @@ class BaseStyleViolation(object):
 
     def node_items(self) -> Tuple[int, int, str]:
         """Returns tuple to match ``flake8`` API format."""
-        line_number = getattr(self._node, 'lineno', 0)
-        column_offset = getattr(self._node, 'col_offset', 0)
-        return line_number, column_offset, self.message()
+        return (*self._location(), self.message())
 
 
 class ASTStyleViolation(BaseStyleViolation):
     """AST based style violations."""
 
-    def __init__(self, node: ast.AST, text: str = None) -> None:
-        """Creates new instance of AST style violation."""
-        super().__init__(node, text=text)
+    _node: ast.AST
+
+    def _location(self) -> Tuple[int, int]:
+        return self._node.lineno, self._node.col_offset
 
 
 class SimpleStyleViolation(BaseStyleViolation):
     """Style violation for cases where there's no AST nodes."""
 
+    _node: None
+
     def __init__(self, node=None, text: str = None) -> None:
         """Creates new instance of simple style violation."""
         super().__init__(node, text=text)
+
+
+class TokenStyleViolation(BaseStyleViolation):
+    """Style violation for ``tokenize`` errors."""
+
+    _node: tokenize.TokenInfo
+
+    def _location(self) -> Tuple[int, int]:
+        return self._node.start