+++ wemake_python_styleguide/visitors/presets/tokens.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+
+from wemake_python_styleguide.visitors.tokenize.wrong_primitives import (
+    WrongPrimitivesVisitor,
+)
+
+TOKENS_PRESET = (
+    WrongPrimitivesVisitor,
+)